Added a Content layout
There is overlap between Content and ContentIndex - possibly look at the possibility of Content inheriting and being the slot in ContentIndex, but probably not (in case we want to change Index in future, very possible, and I think it would be even more clumsy than having these 2 duplicates, cause you'd essentially have to add more __layout files to each folder. So yes, this is the best way).