Layout Infrastructure (#222)
* setup redirects for old naming system

* Revert "setup redirects for old naming system"

This reverts commit 265f53db5ca2b30aed76bfe3bc18685a4d2b98da.

* Modified top-level __layout

Only contains the header, footer and crumbs. Everything is wrapped in the "content" div

* Added a FrontPage layout

This currently duplicates what was happening before, but can now be easily modified.

* Added ContentIndex layout

This layout will handle listed pages (essentially Learn and Explore)

* Added a Content layout

There is overlap between Content and ContentIndex - possibly look at the possibility of Content inheriting and being the slot in ContentIndex, but probably not (in case we want to change Index in future, very possible, and I think it would be even more clumsy than having these 2 duplicates, cause you'd essentially have to add more __layout files to each folder. So yes, this is the best way).

* Lower case layouts

* typo correction

* Making reference index look how it used to

* Making installation index look how it used to

* remove extraneous styles

* move svelte indices to .svx

* small refactor

* change name of related to blank

* use contentindex on front page

* simplify layouts

* use cases that git respects

git is dumb

Co-authored-by: James Bradbury <me@jamesbradbury.net>
Co-authored-by: James Bradbury <github@jamesbradbury.net>